Improved creation of new procedure retval types and retval procedure selectors (#3281)
* Initial

* Manual cherry-pick

* Added check that all used RetvalProcedure ancestors are registered

* Fix all requested changes
