+++ src/main/java/net/mcreator/element/converter/fv32/ItemDispenseBehaviorToItemExtensionConverter.java
@@ -25,7 +25,7 @@ import net.mcreator.element.GeneratableElement;
 import net.mcreator.element.ModElementType;
 import net.mcreator.element.converter.IConverter;
 import net.mcreator.element.parts.MItemBlock;
-import net.mcreator.element.parts.Procedure;
+import net.mcreator.element.parts.procedure.Procedure;
 import net.mcreator.element.types.ItemExtension;
 import net.mcreator.workspace.Workspace;
 import net.mcreator.workspace.elements.FolderElement;
