+++ src/main/java/net/mcreator/element/parts/gui/InputSlot.java
@@ -19,7 +19,7 @@
 package net.mcreator.element.parts.gui;
 
 import net.mcreator.element.parts.MItemBlock;
-import net.mcreator.element.parts.Procedure;
+import net.mcreator.element.parts.procedure.Procedure;
 
 import java.awt.*;
