+++ src/main/java/net/mcreator/element/parts/gui/OutputSlot.java
@@ -18,7 +18,7 @@
 
 package net.mcreator.element.parts.gui;
 
-import net.mcreator.element.parts.Procedure;
+import net.mcreator.element.parts.procedure.Procedure;
 
 import java.awt.*;
