+++ src/main/java/net/mcreator/ui/modgui/ProcedureGUI.java
@@ -22,7 +22,7 @@ import net.mcreator.blockly.BlocklyCompileNote;
 import net.mcreator.blockly.data.*;
 import net.mcreator.blockly.java.BlocklyToProcedure;
 import net.mcreator.element.GeneratableElement;
-import net.mcreator.element.parts.Procedure;
+import net.mcreator.element.parts.procedure.Procedure;
 import net.mcreator.element.parts.gui.GUIComponent;
 import net.mcreator.element.types.Command;
 import net.mcreator.element.types.GUI;
